requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ML & CV
torch>=2.9.0
torchvision>=0.19.0
mediapipe>=0.10.9
opencv-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
